backups/政务平台_20260426_140322/phase1_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
requests>=2.31.0
websocket-client>=1.7.0
